requirements.txt
bcrypt==4.0.1
click==8.1.7
cmake==3.27.7
colorama==0.4.6
joblib==1.3.2
mysql-connector-python==8.2.0
numpy==1.26.2
opencv-python==4.8.1.78
packaging==23.2
ply==3.11
protobuf==4.21.12
PyQt5Designer==5.14.1
PyQt6==6.6.0
pyqt6-plugins==6.4.2.2.3
PyQt6-Qt6==6.6.0
PyQt6-sip==13.6.0
pyqt6-tools==6.4.2.3.3
python-dotenv==1.0.0
qt6-applications==6.4.3.2.3
qt6-tools==6.4.3.1.3
scikit-learn==1.3.2
scipy==1.11.4
sip==6.7.12
threadpoolctl==3.2.0
